src/requirements.txt
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==11.3.0
pymongo==4.15.2
pytesseract==0.3.13
Werkzeug==3.1.3
